second commit

+++ RENATINHO 1 (V-4)/renatinhoSoftware/mpu6050.h
@@ -136,15 +136,15 @@ void mpu_reset() {
 }
 
 /* Retorna o angulo de giro em Z */
-float getAngleX() {
+int getAngleX() {
   return roll;
 }
 
-float getAngleY() {
+int getAngleY() {
   return pitch;
 }
 
-float getAngleZ() {
+int getAngleZ() {
   return yaw;
 }
